feat(zeitrahmen): Add new zeitrahmen field for CLI
In ImpfterminService, rename `zeitspanne` to `zeitrahmen` and change its
format to be more user-friendly. Also, make it possible to configure
this `zeitrahmen` via the `.zeitrahmen` field of `kontaktdaten.json`
and via the CLI. As before, one can construct a fully-featured
`kontaktdaten.json` via the CLI.

This commit also adapts the GUI to pass the new format to the
ImpfterminService. However, it will need further changes to migrate
from `zeitspanne.json` to `kontaktdaten.json`, so that GUI and CLI
use the same storage format.

Signed-off-by: Sebastian Hasler <sebastian.hasler@gmx.net>